Add existingWhitespace::merge_into(…)
